Made error checking much more user configurable.
The choice of which gphoto2 return values should raise an exception is
very program dependent, I've added an 'error_severity' dict which the
user can use to assign a different severity to selected return values.
There is also an overall 'error_exception' value that sets how severe an
error has to be before an exception is raised.