Use "friendly" name in "see also" docstring
Each GPhoto2 struct has a name like _Xyz which is typedeffed to Xyz.
SWIG requires the _Xyz name to extend the struct, but it's nicer to use
the Xyz form in the docstrings. Hence we now pass both to the macro.
(It's a pity SWIG can't derive _Xyz from Xyz in a macro - it can do it
in a typemap.)